// src/modules/@inventory/constant/enums.ts
export enum HouseTypes {}

export enum ProductTypeEnum {
  Inventory = "Inventory",
  Casual = "Casual",
}

export enum StateEnum {
  Pending = "Pending",
  Success = "Success",
  Failed = "failed",
  Cancelled = "Cancelled",
}

export enum ProcurementEnum {
  Green = "Green",
  Blue = "Blue",
}

export enum InventoryBillUserTypeEnum {
  Staff = "Staff",
  Student = "Student",
  Scholar = "Scholar",
  Other = "Other",
}

export enum ProductCategoryEnum {
  Packet = "Packet",
  Liquid = "Liquid",
  Weight = "Weight",
}

export enum TransactionFlowEnum {
  IN = "IN",
  OUT = "OUT",
}

export enum StoreGroupType {
  Student = "Student",
  Employee = "Employee",
}